Created an InputHandler script which reads inputs from a PS4 controller
This script can be used to read the inputs for any game so I want to save it as a thing I can use again in the future. To go along with it, I've also set up an Input Manager preset which has all of the PS4 controller buttons set up properly. In theory I can use any controller if I create a different preset for it because all modern controllers basically have the same number of buttons and sticks.